requirements.txt
numpy==2.1.3
Pillow==11.0.0
streamlit==1.31.1
tensorflow==2.17.0
tensorflow_intel==2.17.0
